fix int long type mismatch

+++ src/ssl.c
@@ -5703,7 +5703,7 @@ WOLFSSL_EVP_PKEY* wolfSSL_d2i_PrivateKey(int type, WOLFSSL_EVP_PKEY** out,
     }
 
     local->type     = type;
-    local->pkey_sz  = inSz;
+    local->pkey_sz  = (int)inSz;
     local->pkey.ptr = (char*)XMALLOC(inSz, NULL, DYNAMIC_TYPE_PUBLIC_KEY);
     if (local->pkey.ptr == NULL) {
         wolfSSL_EVP_PKEY_free(local);
